Refactor payout externality checks to use getEffectivePayoutMode for improved clarity and consistency across components.
