+++ apps/web/app/(ee)/partners.dub.co/(dashboard)/payouts/payout-table.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { isPayoutExternal } from "@/lib/api/payouts/is-payout-external-for-program";
+import { getEffectivePayoutMode } from "@/lib/api/payouts/get-effective-payout-mode";
 import { INVOICE_AVAILABLE_PAYOUT_STATUSES } from "@/lib/partners/constants";
 import usePartnerPayouts from "@/lib/swr/use-partner-payouts";
 import usePartnerPayoutsCount from "@/lib/swr/use-partner-payouts-count";
@@ -229,9 +229,9 @@ function AmountRowItem({ payout }: { payout: PartnerPayoutResponse }) {
     return null;
   }
 
-  const isExternal = isPayoutExternal({
+  const payoutMode = getEffectivePayoutMode({
     payoutMode: payout.program.payoutMode,
-    payoutsEnabledAt: partner?.payoutsEnabledAt,
+    payoutsEnabledAt: partner?.payoutsEnabledAt ?? null,
   });
 
   const display = currencyFormatter(payout.amount / 100);
@@ -262,7 +262,7 @@ function AmountRowItem({ payout }: { payout: PartnerPayoutResponse }) {
   return (
     <div className="flex items-center gap-1.5">
       {display}
-      {isExternal && (
+      {payoutMode === "external" && (
         <Tooltip
           content={
             payout.status === PayoutStatus.pending
