+++ apps/web/lib/api/payouts/get-effective-payout-mode.ts
@@ -1,26 +1,5 @@
 import { PayoutMode, ProgramPayoutMode } from "@prisma/client";
 
-interface IsPayoutExternalForProgramProps {
-  payoutMode: ProgramPayoutMode | null;
-  payoutsEnabledAt: Date | null;
-}
-
-export function isPayoutExternal({
-  payoutMode,
-  payoutsEnabledAt,
-}: IsPayoutExternalForProgramProps): boolean {
-  switch (payoutMode) {
-    case "internal":
-      return false;
-    case "external":
-      return true;
-    case "hybrid":
-      return payoutsEnabledAt === null;
-    default:
-      throw new Error(`Invalid payout mode: ${payoutMode}`);
-  }
-}
-
 export function getEffectivePayoutMode({
   payoutMode,
   payoutsEnabledAt,
